Redmine6 compatibility
