--- lib/gemavatar_hooks.rb
@@ -1,20 +1,0 @@
-#    This file is part of Gemavatar.
-#
-#    Gemavatar is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    Gemavatar is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with Gemavatar.  If not, see <http://www.gnu.org/licenses/>.
-
-module GemAvatarPlugin
-    class GemAvatarHooks < Redmine::Hook::ViewListener
-        render_on :view_my_account, :partial => 'hooks/reload'
-    end
-end
